backend/requirements.txt
flask
flask-cors
mysql-connector-python
pdfplumber
python-docx
werkzeug
gunicorn
